fix: rename BookingSummary folder, add styled.ts
