+++ source/screens/bookingScreens/BookingSummary/BookingSummary.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import styled from "styled-components/native";
 import { BookingItem } from "../../../types/BookingTypes";
 import icons from "../../../helpers/Icons";
 import { Text } from "../../../components/atoms";
@@ -8,23 +7,7 @@ import {
   AddressCard,
   CharacterCard,
 } from "../../../components/molecules";
-
-const Container = styled.ScrollView`
-  height: 80%;
-  margin-top: 0px;
-  padding-top: 0px;
-  padding-left: 16px;
-  padding-right: 16px;
-  background-color: ${({ theme }) => theme.colors.neutrals[6]};
-`;
-
-const TitleWrapper = styled.View`
-  margin-bottom: 15px;
-`;
-
-const Title = styled(Text)`
-  margin: 15px 0px;
-`;
+import { Container, TitleWrapper, Title } from "./styled";
 
 interface BookingSummaryProps {
   route: any;
